Added show_AOD_bin function to AOD module.
The function takes in a veltable and a bin number and shows the data and
highlights the selected bin in one panel, while showing the flam-relflux
plot with best AOD fit and illurtrative high-N-low-fc and vice versa  in
the same panel. No fancy computations, but good for understanding what
is happening.