New minor version: Added keyboard controls
Bumped version number from 0.1.5 to 0.2 on account of those keyboard
controls that I added to the thing.

Also found out that there was a bug introduced in Matplotlib that
renders SpecView only half usable, so Matplotlib 3.1.x for now, until
there is a solution. Unfortunately, it the issue was just postponed to
the Matplotlib 3.5 milestone (current version is 3.4.2).

Also did a BUGFIX, so now, skylines work without having to manually move
the folder around anymore, that's nice.

Kthxbai!